Move powerbutton to its own class
